requirements.txt
beautifulsoup4
fastapi
mysqlclient
python-dotenv
requests
slowapi
starlette
uvicorn
pymongo
